# core/api/openapi.yaml
openapi: 3.0.3
info:
  title: Ballet Core API
  version: v1
  description: |
    Stateless REST API of Ballet Core. All operations require an OIDC access
    token (`Authorization: Bearer …`) and are authorized through role
    bindings. Errors use the `Error` schema. Updates use optimistic
    concurrency through `version`.
servers:
  - url: /
security:
  - bearer: []
paths:
...
  /api/v1/organizations/{organization}/knowledge/entries:
    parameters: [{ $ref: "#/components/parameters/Organization" }]
    get:
      operationId: listKnowledgeEntries
      tags: [knowledge]
      summary: Entries of the organization's knowledge space (served by Knowledge through Core)
      parameters:
        - { name: kind, in: query, schema: { $ref: "#/components/schemas/KnowledgeKind" } }
        - { name: project, in: query, schema: { type: string }, description: Project key }
        - { name: item, in: query, schema: { type: string }, description: Linked tracker item key }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/KnowledgeEntryList" } } } }
        default: { $ref: "#/components/responses/Error" }
    post:
      operationId: createKnowledgeEntry
      tags: [knowledge]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CreateKnowledgeEntry" } } }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/KnowledgeEntry" } } } }
        default: { $ref: "#/components/responses/Error" }
  /api/v1/organizations/{organization}/knowledge/search:
    parameters: [{ $ref: "#/components/parameters/Organization" }]
    get:
      operationId: searchKnowledge
      tags: [knowledge]
      summary: Hybrid full-text and semantic search in the organization's knowledge space
      parameters:
        - { name: q, in: query, required: true, schema: { type: string } }
        - { name: kind, in: query, schema: { $ref: "#/components/schemas/KnowledgeKind" } }
        - { name: project, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 50, default: 20 } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/KnowledgeSearchResult" } } } }
        default: { $ref: "#/components/responses/Error" }
  /api/v1/organizations/{organization}/knowledge/entries/{entry}:
    parameters: [{ $ref: "#/components/parameters/Organization" }, { $ref: "#/components/parameters/Entry" }]
    get:
      operationId: getKnowledgeEntry
      tags: [knowledge]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/KnowledgeEntry" } } } }
        default: { $ref: "#/components/responses/Error" }
    patch:
      operationId: updateKnowledgeEntry
      tags: [knowledge]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/UpdateKnowledgeEntry" } } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/KnowledgeEntry" } } } }
        default: { $ref: "#/components/responses/Error" }
  /api/v1/organizations/{organization}/knowledge/entries/{entry}/versions:
    parameters: [{ $ref: "#/components/parameters/Organization" }, { $ref: "#/components/parameters/Entry" }]
    get:
      operationId: listKnowledgeVersions
      tags: [knowledge]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/KnowledgeVersionList" } } } }
        default: { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    Organization: { name: organization, in: path, required: true, schema: { type: string }, description: Organization key }
...
    Entry: { name: entry, in: path, required: true, schema: { type: string }, description: Knowledge entry ID }
...
  responses:
    Error:
      description: Error
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  schemas:
    Error:
      type: object
      additionalProperties: false
      required: [error, message]
      properties:
        error:
          type: string
          enum: [invalid_argument, unauthenticated, forbidden, not_found, already_exists, conflict, internal]
        message: { type: string }
...
    KnowledgeKind: { type: string, enum: [document, decision, note, debt] }
    KnowledgeEntry:
      type: object
      additionalProperties: false
      required: [id, kind, title, body, projects, items, version, created_by, updated_by, created_at, updated_at]
      properties:
        id: { type: string }
        kind: { $ref: "#/components/schemas/KnowledgeKind" }
        title: { type: string }
        body: { type: string, description: Markdown }
        projects: { type: array, items: { type: string } }
        items: { type: array, items: { type: string }, description: Linked tracker item keys }
        version: { type: integer, format: int64 }
        created_by: { type: string }
        updated_by: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    KnowledgeEntryList:
      type: object
      additionalProperties: false
      required: [items]
      properties: { items: { type: array, items: { $ref: "#/components/schemas/KnowledgeEntry" } } }
    CreateKnowledgeEntry:
      type: object
      additionalProperties: false
      required: [kind, title]
      properties:
        kind: { $ref: "#/components/schemas/KnowledgeKind" }
        title: { type: string }
        body: { type: string }
        projects: { type: array, items: { type: string } }
        items: { type: array, items: { type: string } }
    UpdateKnowledgeEntry:
      type: object
      additionalProperties: false
      required: [version]
      properties:
        version: { type: integer, format: int64 }
        kind: { $ref: "#/components/schemas/KnowledgeKind" }
        title: { type: string }
        body: { type: string }
        projects: { type: array, items: { type: string } }
        items: { type: array, items: { type: string } }
    KnowledgeVersion:
      type: object
      additionalProperties: false
      required: [version, title, body, author, created_at]
      properties:
        version: { type: integer, format: int64 }
        title: { type: string }
        body: { type: string }
        author: { type: string }
        created_at: { type: string, format: date-time }
    KnowledgeVersionList:
      type: object
      additionalProperties: false
      required: [items]
      properties: { items: { type: array, items: { $ref: "#/components/schemas/KnowledgeVersion" } } }
    KnowledgeSearchHit:
      type: object
      additionalProperties: false
      required: [entry, score]
      properties:
        entry: { $ref: "#/components/schemas/KnowledgeEntry" }
        score: { type: number, description: Fused rank score; higher is better }
    KnowledgeSearchResult:
      type: object
      additionalProperties: false
      required: [items]
      properties: { items: { type: array, items: { $ref: "#/components/schemas/KnowledgeSearchHit" } } }
...